Add multi colon logic
